feat: enhance party synchronization by adding sync command handling, provider change support, and updating data types for startAt

+++ src/main/java/com/riyura/backend/modules/party/dto/PartyStateResponse.java
@@ -17,7 +17,7 @@ public class PartyStateResponse {
     private int seasonNo;
     private int episodeNo;
     private String providerId;
-    private int startAt;
+    private double startAt;
     private long partyStartedAt;
     private boolean strictSync;
     private List<String> participantIds;
